Fix the controller crashing issue
If no spec.selector.matchingLabels is set, the application controller
crashes due to a nil pointer. This commit fixes the issue by checking if
selector is nil before getting the matching labels from it.

Per the API doc for the LabelSelector:
https://github.com/kubernetes/apimachinery/blob/76330795f827e44ffc47f4fdc8e0d39872bead8f/pkg/apis/meta/v1/types.go#L1060-L1061.
If spec.selector is nil, it should return no objects.
If spec.selector is empty, it should match all objects.